+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/PassScreen.kt
@@ -22,9 +22,9 @@ import kotlinx.coroutines.withContext
 import nz.eloque.foss_wallet.R
 import nz.eloque.foss_wallet.model.Pass
 import nz.eloque.foss_wallet.ui.WalletScaffold
-import nz.eloque.foss_wallet.ui.components.pass_view.PassShareButton
-import nz.eloque.foss_wallet.ui.components.pass_view.PassView
-import nz.eloque.foss_wallet.ui.wallet.PassViewModel
+import nz.eloque.foss_wallet.ui.view.pass.PassShareButton
+import nz.eloque.foss_wallet.ui.view.pass.PassView
+import nz.eloque.foss_wallet.ui.view.wallet.PassViewModel
 import java.util.Locale
 
 @OptIn(ExperimentalMaterial3Api::class)
