+++ app/src/main/java/nz/eloque/foss_wallet/ui/view/pass/PassView.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.components.pass_view
+package nz.eloque.foss_wallet.ui.view.pass
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -18,7 +18,8 @@ import androidx.compose.ui.unit.dp
 import nz.eloque.foss_wallet.model.Pass
 import nz.eloque.foss_wallet.model.PassType
 import nz.eloque.foss_wallet.model.field.PassContent
-import nz.eloque.foss_wallet.ui.components.card.PassCard
+import nz.eloque.foss_wallet.model.field.PassField
+import nz.eloque.foss_wallet.ui.card.PassCard
 import java.time.Instant
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -74,20 +75,20 @@ private fun PassPreview() {
     ).also {
         it.relevantDate = 1800000000L
         it.headerFields = mutableListOf(
-            nz.eloque.foss_wallet.model.field.PassField("block", "Block", PassContent.Plain("S1")),
-            nz.eloque.foss_wallet.model.field.PassField("seat", "Seat", PassContent.Plain("47")),
+            PassField("block", "Block", PassContent.Plain("S1")),
+            PassField("seat", "Seat", PassContent.Plain("47")),
         )
         it.primaryFields = mutableListOf(
-            nz.eloque.foss_wallet.model.field.PassField("name", "Name", PassContent.Plain("Max Mustermann")),
-            nz.eloque.foss_wallet.model.field.PassField("seat", "Seat", PassContent.Plain("47")),
+            PassField("name", "Name", PassContent.Plain("Max Mustermann")),
+            PassField("seat", "Seat", PassContent.Plain("47")),
         )
         it.auxiliaryFields = mutableListOf(
-            nz.eloque.foss_wallet.model.field.PassField("block", "Block", PassContent.Plain("S1 | Gegengerade")),
-            nz.eloque.foss_wallet.model.field.PassField("seat", "Seat", PassContent.Plain("36E")),
+            PassField("block", "Block", PassContent.Plain("S1 | Gegengerade")),
+            PassField("seat", "Seat", PassContent.Plain("36E")),
         )
         it.secondaryFields = mutableListOf(
-            nz.eloque.foss_wallet.model.field.PassField("data1", "data1", PassContent.Plain("Longer Value here i guess")),
-            nz.eloque.foss_wallet.model.field.PassField("data2", "data2", PassContent.Plain("Shorter Value")),
+            PassField("data1", "data1", PassContent.Plain("Longer Value here i guess")),
+            PassField("data2", "data2", PassContent.Plain("Shorter Value")),
         )
     }
     PassView(pass)
